requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
Pillow>=10.0.0
django-filter>=23.0
markdown>=3.5

